Factor out saving map to helper module
